+++ src/main/java/com/github/edg_thexu/better_experience/client/event/ClientModEvent.java
@@ -14,7 +14,7 @@ import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.EntityRenderersEvent;
 import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
 
-@EventBusSubscriber(modid = Better_experience.MODID, bus = EventBusSubscriber.Bus.MOD,value = Dist.CLIENT)
+@EventBusSubscriber(modid = Better_experience.MODID,value = Dist.CLIENT)
 public class ClientModEvent {
 
     @SubscribeEvent