+++ src/main/java/com/github/edg_thexu/better_experience/item/DebugItem.java
@@ -1,8 +1,5 @@
 package com.github.edg_thexu.better_experience.item;
 
-import com.github.edg_thexu.better_experience.client.gui.editor.ObjectEditorScreen;
-import net.minecraft.client.Minecraft;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.player.Player;
@@ -21,10 +18,11 @@ public class DebugItem extends Item {
     public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand usedHand) {
 
         if(level.isClientSide()){
-            Minecraft.getInstance().setScreen(new ObjectEditorScreen(Component.literal("Editor") ));
-            return InteractionResultHolder.success(player.getItemInHand(usedHand));
-
+//            Minecraft.getInstance().setScreen(new ObjectEditorScreen(Component.literal("Editor") ));
+//            return InteractionResultHolder.success(player.getItemInHand(usedHand));
         }
         return super.use(level, player, usedHand);
     }
+
+
 }