+++ src/main/java/com/github/edg_thexu/better_experience/mixin/FishingHookMixin.java
@@ -26,7 +26,8 @@ import java.util.List;
 @Mixin(FishingHook.class)
 public abstract class FishingHookMixin implements IFishingHook, SelfGetter<FishingHook> {
 
-    @Shadow private int nibble;
+    @Shadow
+    public int nibble;
 
     @Shadow public abstract int retrieve(ItemStack stack);