+++ src/main/java/com/github/edg_thexu/better_experience/mixin/ProjectileWeaponItemMixin.java
@@ -6,7 +6,6 @@ import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ProjectileWeaponItem;
-import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
@@ -26,6 +25,6 @@ public class ProjectileWeaponItemMixin {
         if(CommonConfig.INFINITE_AMMO.get() &&  ammo.getCount() > CommonConfig.INFINITE_AMMO_STACK_SIZE.get()){
             return 0;
         }
-        return EnchantmentHelper.processAmmoUse(level, weapon, ammo, count);
+        return original.call(level, weapon, ammo, count);
     }
 }