+++ src/main/java/com/github/edg_thexu/better_experience/registries/itemmatcher/ItemStackWrapper.java
@@ -2,24 +2,20 @@ package com.github.edg_thexu.better_experience.registries.itemmatcher;
 
 import com.github.edg_thexu.better_experience.registries.itemmatcher.variant.IdCountMatcher;
 import com.mojang.serialization.Codec;
-import com.mojang.serialization.codecs.RecordCodecBuilder;
 import mezz.jei.api.ingredients.ITypedIngredient;
-import mezz.jei.common.Internal;
-import mezz.jei.common.codecs.TypedIngredientCodecs;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-
-import java.util.Optional;
 
 public class ItemStackWrapper {
 
     ITypedIngredient<?> ingredient;
     IItemMatcher matcher;
 
-    public static Codec<ItemStackWrapper> CODEC = RecordCodecBuilder.create(instance -> instance.group(
-            TypedIngredientCodecs.getIngredientCodec(Internal.getJeiRuntime().getIngredientManager()).codec().fieldOf("ingredients").forGetter(ItemStackWrapper::ingredients),
-            IItemMatcher.TYPE_CODEC.optionalFieldOf("matcher").forGetter(i->Optional.ofNullable(i.matcher))
-    ).apply(instance, (ingredients, matcher)-> new ItemStackWrapper(ingredients, matcher.orElse(IdCountMatcher.INSTANCE))));
+    public static Codec<ItemStackWrapper> CODEC = null;
+//    public static Codec<ItemStackWrapper> CODEC = RecordCodecBuilder.create(instance -> instance.group(
+//            TypedIngredientCodecs.getIngredientCodec(Internal.getJeiRuntime().getIngredientManager()).codec().fieldOf("ingredients").forGetter(ItemStackWrapper::ingredients),
+//            IItemMatcher.TYPE_CODEC.optionalFieldOf("matcher").forGetter(i->Optional.ofNullable(i.matcher))
+//    ).apply(instance, (ingredients, matcher)-> new ItemStackWrapper(ingredients, matcher.orElse(IdCountMatcher.INSTANCE))));
 
     public ItemStackWrapper(ITypedIngredient<?> ingredient, IItemMatcher matcher){
         this.ingredient = ingredient;