+++ src/main/java/com/github/edg_thexu/better_experience/registries/recipehandler/IRecipeHandler.java
@@ -3,15 +3,11 @@ package com.github.edg_thexu.better_experience.registries.recipehandler;
 import com.github.edg_thexu.better_experience.intergration.jei.JeiRegistries;
 import com.github.edg_thexu.better_experience.registries.recipehandler.visitor.IRecipeHandlerVisitor;
 import com.mojang.serialization.Codec;
-import mezz.jei.library.gui.recipes.RecipeLayout;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.api.distmarker.OnlyIn;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
@@ -25,10 +21,10 @@ public interface IRecipeHandler<V> {
     /**
      * 初始化时使用，但是工厂方法可以平替
      */
-    @OnlyIn(Dist.CLIENT)
-    default void init(RecipeLayout layout, @Nullable Object recipe){
-
-    }
+//    @OnlyIn(Dist.CLIENT)
+//    default void init(RecipeLayout layout, @Nullable Object recipe){
+//
+//    }
 
     /**
      * 生成每个原料的迭代器