+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-## Fix
-- Compatible with some mod witch change ammo consumption. (Such as lovely-sparkle-pieces) 
+## Delete
+- Remove ingredient fetching feature related with jei